requirements.txt
boto3==1.9.35
botocore==1.12.35
certifi==2018.8.24
chardet==3.0.4
Click==7.0
DBUtils==1.3
docutils==0.14
ffmpy==0.2.2
Flask==1.0.2
futures==3.2.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mysql==0.0.2
MySQL-python==1.2.5
nose==1.3.7
pydevd==1.4.0
python-dateutil==2.7.5
redis==2.10.6
requests==2.19.1
s3transfer==0.1.13
selenium==3.14.1
six==1.11.0
threadpool==1.3.2
urllib3==1.23
Werkzeug==0.14.1
